Settings: Enable App Ops
This patch enables the basic aosp App Ops and is 100% standalone (meaning it needs no other changes to any other projects). It includes all of the minor fixes needed to get this running in Android Marshmallow. Thanks to @MrApocalypse for the extra help debugging the crashes.

BenzoEdit:
Made it work with 7.0

Change-Id: I0f8e37a1fa0fe58bc8e9a62e88a5648fe52a4b4e
Signed-off-by: Chet Kener <Cl3Kener@gmail.com>
Signed-off-by: mydongistiny <jaysonedson@gmail.com>